+++ changelog.md
@@ -1,6 +1,8 @@
 # 0.1.8
 - Fixed
   - Version Checker updates failing if there was a certificate error (specifically sc2mafia.com).
+  - If a file failed to extract, the mod install wouldn't fail and you'd have to check the log to know some file was missed.
+    - Now, a dialog will appear with the error message as well as buttons to open the folders for manual install.
 
 # 0.1.7
 - Added